+++ packages/core-gatsby-ui/src/docs-components/text-color-intents-showcase.treat.ts
@@ -1,5 +1,5 @@
 import { Theme } from '@newrade/core-react-ui';
-import capsize from 'capsize';
+import capsize from '@capsizecss/core';
 import { style } from 'treat';
 
 export const styles = {