+++ packages/core-react-ui/src/utilities/typography-v2.utilities.ts
@@ -2,94 +2,91 @@ import {
   CapsizeTextStyle,
   TextStyle,
   TextVariantStyles,
-  Titles,
-  Typography,
+  TitlesV2,
+  TypographyV2,
   VIEWPORT,
 } from '@newrade/core-design-system';
-import { defaultFontVarNames, defaultFontVars } from '../default-theme/default-typography';
-import { CSSTypography } from '../design-system';
+import { CSSTypographyV2 } from '../design-system';
 import { getCSSFonts, getCSSFontsObject } from './font.utilities';
-import { createCSSCapsizeTextStyle } from './text-capsize.utilities';
+import { createCSSCapsizeTextStyleV2 } from './text-capsize.utilities';
 import { createCSSTextStyle } from './text.utilities';
 import { keys } from './utilities';
 
-export function getCSSTypography({
+export function getCSSTypographyV2({
   baseFontSize,
   fonts,
   titles,
   headings,
   paragraphs,
   labels,
-}: Typography & { baseFontSize: number }): CSSTypography {
-  const titlesStyles = createCSSVariantTextStyles({ variant: titles, baseFontSize });
-  const headingsStyles = createCSSVariantTextStyles({ variant: headings, baseFontSize });
-  const paragraphsStyles = createCSSVariantTextStyles({ variant: paragraphs, baseFontSize });
-  const labelsStyles = createCSSVariantTextStyles({ variant: labels, baseFontSize });
+}: TypographyV2 & { baseFontSize: number }): CSSTypographyV2 {
+  const titlesStyles = createCSSVariantTextStylesV2({ variant: titles, baseFontSize });
+  const headingsStyles = createCSSVariantTextStylesV2({ variant: headings, baseFontSize });
+  const paragraphsStyles = createCSSVariantTextStylesV2({ variant: paragraphs, baseFontSize });
+  const labelsStyles = createCSSVariantTextStylesV2({ variant: labels, baseFontSize });
 
   return {
     fonts: {
       ...getCSSFontsObject(fonts),
-      var: defaultFontVars,
-      varNames: defaultFontVarNames,
     },
     titles: {
       font: getCSSFonts(titles.font ? titles.font : fonts.sans), // fallback to sans font
-      ...(titlesStyles as Typography<string>['titles']),
+      ...(titlesStyles as TypographyV2<string>['titles']),
     },
     headings: {
       font: getCSSFonts(titles.font ? titles.font : fonts.sans), // fallback to sans font
-      ...(headingsStyles as Typography<string>['headings']),
+      ...(headingsStyles as TypographyV2<string>['headings']),
     },
     paragraphs: {
       font: getCSSFonts(titles.font ? titles.font : fonts.sans), // fallback to sans font
-      ...(paragraphsStyles as Typography<string>['paragraphs']),
+      ...(paragraphsStyles as TypographyV2<string>['paragraphs']),
     },
     labels: {
       font: getCSSFonts(titles.font ? titles.font : fonts.sans), // fallback to sans font
-      ...(labelsStyles as Typography<string>['labels']),
+      ...(labelsStyles as TypographyV2<string>['labels']),
     },
   };
 }
 
-function createCSSVariantTextStyles({
+function createCSSVariantTextStylesV2({
   variant,
   baseFontSize,
 }: {
   variant:
-    | Typography['titles']
-    | Typography['headings']
-    | Typography['paragraphs']
-    | Typography['labels'];
+    | TypographyV2['titles']
+    | TypographyV2['headings']
+    | TypographyV2['paragraphs']
+    | TypographyV2['labels'];
   baseFontSize: number;
 }) {
   const parentTextStyles = keys(variant)
     .filter((viewport) => !Object.values(VIEWPORT).includes(viewport as VIEWPORT))
     .reduce((previous, prop) => {
-      const styles = variant[prop] as Omit<Typography['titles'], 'desktop' | 'tablet' | 'mobile'>;
+      const styles = variant[prop] as Omit<TypographyV2['titles'], 'desktop' | 'tablet' | 'mobile'>;
       (previous as any)[prop] = styles;
       return previous;
-    }, {} as Omit<Typography['titles'], 'desktop' | 'tablet' | 'mobile'>);
+    }, {} as Omit<TypographyV2['titles'], 'desktop' | 'tablet' | 'mobile'>);
 
   const variantStyles = keys(variant)
     .filter((viewport) => Object.values(VIEWPORT).includes(viewport as VIEWPORT))
     .reduce((previous, viewport) => {
-      const styles = variant[viewport] as Titles;
+      const styles = variant[viewport] as TitlesV2;
       const viewportStyle = keys(styles).reduce((previous, title) => {
         // merge styles set on parent with specific styles
         const textStyle: TextStyle & CapsizeTextStyle = {
           ...parentTextStyles,
           ...styles[title],
         };
-        const cssTextStyle = createCSSCapsizeTextStyle({ ...textStyle, baseFontSize });
+        const cssTextStyle = createCSSCapsizeTextStyleV2({ ...textStyle, baseFontSize });
         (previous as any)[title] = cssTextStyle;
         return previous;
-      }, {} as Titles<string>);
+      }, {} as TitlesV2<string>);
       (previous as any)[viewport] = viewportStyle;
       return previous;
-    }, {} as Typography<string>['titles'] | Typography<string>['headings'] | Typography<string>['paragraphs'] | Typography<string>['labels']);
+    }, {} as TypographyV2<string>['titles'] | TypographyV2<string>['headings'] | TypographyV2<string>['paragraphs'] | TypographyV2<string>['labels']);
 
-  if (parentTextStyles && (parentTextStyles as Typography['paragraphs']).styles) {
-    const variantStylesStyles = (parentTextStyles as Typography['paragraphs']).styles;
+  if (parentTextStyles && (parentTextStyles as TypographyV2['paragraphs']).styles) {
+    const variantStylesStyles = (parentTextStyles as TypographyV2['paragraphs']).styles;
     return {
       ...variantStyles,
       styles: keys(variantStylesStyles).reduce((previous, current) => {